+++ lib/widgets/common/app_bar/app_bar_subtitle.dart
@@ -1,4 +1,3 @@
-import 'dart:ui';
 
 import 'package:aves/model/source/collection_source.dart';
 import 'package:aves/model/source/events.dart';