+++ lib/widgets/common/identity/highlight_title.dart
@@ -1,4 +1,3 @@
-import 'dart:ui';
 
 import 'package:aves/model/settings/settings.dart';
 import 'package:aves/theme/colors.dart';