+++ lib/widgets/common/search/page.dart
@@ -1,4 +1,3 @@
-import 'dart:ui';
 
 import 'package:aves/theme/durations.dart';
 import 'package:aves/theme/themes.dart';